+++ bika/lims/browser/analysisrequest/add2.py
@@ -506,7 +506,8 @@ class AnalysisRequestAddView(BrowserView):
 
         for brain in services:
             category = brain.getCategoryTitle
-            analyses[category].append(brain)
+            if category in analyses:
+                analyses[category].append(brain)
         return analyses
 
     @cache(cache_key)
@@ -1084,7 +1085,7 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
 
         # client
         client = self.get_client()
-        client_uid = api.get_uid(client) if client else ""
+        client_uid = client and api.get_uid(client) or ""
 
         # sample matrix
         sample_matrix = obj.getSampleMatrix()
@@ -1647,7 +1648,9 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
         # extract records from request
         records = self.get_records()
 
-        errors = {}
+        fielderrors = {}
+        errors = {"message": "", "fielderrors": {}}
+
         attachments = {}
         valid_records = []
 
@@ -1702,7 +1705,7 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
             for field in missing:
                 fieldname = "{}-{}".format(field, n)
                 msg = _("Field '{}' is required".format(field))
-                errors[fieldname] = msg
+                fielderrors[fieldname] = msg
 
             # Selected Analysis UIDs
             selected_analysis_uids = record.get("Analyses", [])
@@ -1758,7 +1761,9 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
             # append the valid record to the list of valid records
             valid_records.append(valid_record)
 
-        if errors:
+        # return immediately with an error response if some field checks failed
+        if fielderrors:
+            errors["fielderrors"] = fielderrors
             return {'errors': errors}
 
         # Process Form
@@ -1774,12 +1779,11 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
             specifications = record.pop("Specifications", {})
 
             # Create the Analysis Request
-            ar = crar(
-                client,
-                self.request,
-                record,
-                specifications=specifications,
-            )
+            try:
+                ar = crar(client, self.request, record, specifications=specifications)
+            except (KeyError, RuntimeError) as e:
+                errors["message"] = e.message
+                return {"errors": errors}
             ARs.append(ar.Title())
 
             _attachments = []