+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog
 
 **Changed**
 
+- #401 PR-2321 AR Add Copy of multiple ARs from different clients raises a Traceback in the background
 
 
 **Fixed**